// core/src/enums/chatroom-types.ts
export enum ChatroomTypes {
  NORMAL_CHATROOM = 0,
  INTRO_CHATROOM = 1,
  EVENT_CHATROOM = 2,
  POLL_CHATROOM = 3,
  HIDDEN_CHATROOM = 4,
  UN_VERIFY_CHATROOM = 5,
  EVENT_PUBLIC_CHATROOM = 6,
  ANNOUNCEMENT_CHATROOM = 7,
  DATE_CHATROOM = 8,
  INTRODUCTION_CHATROOM = 9,
  DIRECT_MESSAGE_CHATROOM = 10,
}